Create Text-Based Game
The final project for MAT-140 was where we had to create a text-based game. Ninja Blade Text Adventure Game. In order to win the game, all items must be retrieved before being caught by Ninja Blade. The text-based game has an inventory and loops prompting a user on what to do next.